requirements.txt
absl-py
anyio
argon2-cffi
argon2-cffi-bindings
asttokens
astunparse
async-lru
attrs
babel
backcall
beautifulsoup4
bleach
Bottleneck
Brotli
brotlicffi
certifi
cffi
charset-normalizer
colorama
comm
contourpy
cycler
debugpy
decorator
defusedxml
exceptiongroup
executing
fastjsonschema
flatbuffers
fonttools
gast
google-pasta
greenlet
grpcio
h11
h5py
httpcore
httpx
idna
imbalanced-learn
importlib_metadata
importlib_resources
ipykernel
ipython
ipywidgets
jedi
Jinja2
joblib
json5
jsonschema
jsonschema-specifications
jupyter
jupyter_client
jupyter-console
jupyter_core
jupyter-events
jupyter-lsp
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
kagglehub
keras
kiwisolver
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mistune
mkl_fft
mkl_random
mkl-service
ml_dtypes
mpi4py
namex
nbclient
nbconvert
nbformat
nest-asyncio
notebook
notebook_shim
numexpr
numpy
opt-einsum
optree
overrides
packaging
pandas
pandocfilters
parso
pickleshare
pillow
pip
platformdirs
prometheus_client
prompt-toolkit
protobuf
psutil
psycopg2
pure-eval
pycparser
Pygments
pyparsing
PyQt6
PyQt6_sip
PySocks
python-dateutil
python-json-logger
pytz
PyYAML
pyzmq
qtconsole
QtPy
referencing
requests
rfc3339-validator
rfc3986-validator
rich
rpds-py
scikit-learn
scipy
seaborn
Send2Trash
setuptools
sip
six
sniffio
soupsieve
SQLAlchemy
stack-data
tensorboard
tensorboard_data_server
tensorflow_cpu
termcolor
terminado
threadpoolctl
tinycss2
tomli
tornado
tqdm
traitlets
typing_extensions
tzdata
unicodedata2
urllib3
wcwidth
webencodings
websocket-client
Werkzeug
wheel
widgetsnbextension
wrapt
zipp
